fixed some bugs in backend

+++ app/dashboard/reports/pregnant/page.tsx
@@ -18,33 +18,41 @@ import {
   SelectTrigger,
   SelectValue,
 } from "@/components/ui/select";
-import { Eye, Edit, Trash2 } from "lucide-react";
+import { Eye, Edit, Trash2, ArrowLeft } from "lucide-react";
 import Link from "next/link";
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger,
+} from "@/components/ui/alert-dialog";
+import { fetchUsersData, PUROKS } from "@/lib/constants";
+import { IPregnant } from "@/lib/models/pregnant";
 import { deletePregnantRecord } from "@/app/actions/pregnant-response";
-
-type PregnantWomanRecord = {
-  _id: string;
-  firstName: string;
-  lastName: string;
-  age: number;
-  assignedStaff: string;
-  address: string;
-};
-
-const purokOptions = Array.from({ length: 10 }, (_, i) => `Purok ${i + 1}`);
-const staffOptions = ["Nurse Garcia", "Midwife Lim", "Dr. Tan"];
 
 export default function PregnantWomenReports() {
   const [lastNameSearch, setLastNameSearch] = useState("");
   const [addressFilter, setAddressFilter] = useState("all");
   const [staffFilter, setStaffFilter] = useState("all");
-  const [records, setRecords] = useState<PregnantWomanRecord[]>([]);
+  const [records, setRecords] = useState<IPregnant[]>([]);
   const [loading, setLoading] = useState(true);
   const [currentPage, setCurrentPage] = useState(1);
   const [error, setError] = useState<string | null>(null);
+  const [staffList, setStaffList] = useState<string[]>([]);
+
   const itemsPerPage = 5;
 
   useEffect(() => {
+    const loadStaffData = async () => {
+      const data = await fetchUsersData();
+      setStaffList(data);
+    };
+    loadStaffData();
     fetchPregnantWomenData();
   }, []);
 
@@ -57,7 +65,7 @@ export default function PregnantWomenReports() {
         throw new Error("Failed to fetch data");
       }
       const data = await response.json();
-      setRecords(data);
+      setRecords(data.data);
     } catch (err) {
       setError(
         "An error occurred while fetching data. Please try again later."
@@ -83,11 +91,29 @@ export default function PregnantWomenReports() {
   const totalPages = Math.ceil(filteredRecords.length / itemsPerPage);
 
   const handleDelete = async (id: string) => {
-    await deletePregnantRecord(id);
+    try {
+      const res = await deletePregnantRecord(id);
+      if (!res?.success) {
+        console.log("Failed to delete");
+      }
+      fetchPregnantWomenData();
+    } catch (error) {}
   };
 
   return (
-    <div className="flex flex-col min-h-screen ">
+    <div className="flex flex-col min-h-screen">
+      <div className="flex items-center justify-between">
+        <Link href="/dashboard/reports/">
+          <Button
+            variant="outline"
+            size="sm"
+            className="flex items-center gap-2"
+          >
+            <ArrowLeft className="w-4 h-4" />
+            Back to Records
+          </Button>
+        </Link>
+      </div>
       <main className="flex-grow p-4 md:p-6 lg:p-8">
         <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-4 gap-4 mb-6">
           <Input
@@ -103,7 +129,7 @@ export default function PregnantWomenReports() {
             </SelectTrigger>
             <SelectContent>
               <SelectItem value="all">All Puroks</SelectItem>
-              {purokOptions.map((purok) => (
+              {PUROKS.map((purok) => (
                 <SelectItem key={purok} value={purok}>
                   {purok}
                 </SelectItem>
@@ -116,7 +142,7 @@ export default function PregnantWomenReports() {
             </SelectTrigger>
             <SelectContent>
               <SelectItem value="all">All Staff</SelectItem>
-              {staffOptions.map((staff) => (
+              {staffList.map((staff) => (
                 <SelectItem key={staff} value={staff}>
                   {staff}
                 </SelectItem>
@@ -160,7 +186,7 @@ export default function PregnantWomenReports() {
                     <TableCell>
                       <div className="flex space-x-2">
                         <Link
-                          href={"/dashboard/reports/pregnant/" + record._id}
+                          href={`/dashboard/reports/pregnant/${record._id}`}
                         >
                           <Button size="sm" variant="outline" className="p-1">
                             <Eye className="h-4 w-4" />
@@ -168,24 +194,41 @@ export default function PregnantWomenReports() {
                           </Button>
                         </Link>
                         <Link
-                          href={
-                            "/dashboard/reports/pregnant/edit/" + record._id
-                          }
+                          href={`/dashboard/reports/pregnant/edit/${record._id}`}
                         >
                           <Button size="sm" variant="outline" className="p-1">
                             <Edit className="h-4 w-4" />
                             <span className="sr-only">Edit</span>
                           </Button>
                         </Link>
-                        <Button
-                          size="sm"
-                          variant="outline"
-                          onClick={() => handleDelete(record._id)}
-                          className="p-1"
-                        >
-                          <Trash2 className="h-4 w-4" />
-                          <span className="sr-only">Delete</span>
-                        </Button>
+                        <AlertDialog>
+                          <AlertDialogTrigger asChild>
+                            <Button size="sm" variant="outline" className="p-1">
+                              <Trash2 className="h-4 w-4" />
+                              <span className="sr-only">Delete</span>
+                            </Button>
+                          </AlertDialogTrigger>
+                          <AlertDialogContent className="bg-white">
+                            <AlertDialogHeader>
+                              <AlertDialogTitle>
+                                Are you sure you want to delete this record?
+                              </AlertDialogTitle>
+                              <AlertDialogDescription>
+                                This action cannot be undone. This will
+                                permanently delete the pregnant woman&apos;s
+                                record.
+                              </AlertDialogDescription>
+                            </AlertDialogHeader>
+                            <AlertDialogFooter>
+                              <AlertDialogCancel>Cancel</AlertDialogCancel>
+                              <AlertDialogAction
+                                onClick={() => handleDelete(record._id)}
+                              >
+                                Delete
+                              </AlertDialogAction>
+                            </AlertDialogFooter>
+                          </AlertDialogContent>
+                        </AlertDialog>
                       </div>
                     </TableCell>
                   </TableRow>
@@ -225,5 +268,5 @@ export default function PregnantWomenReports() {
 }
 
 function SkeletonTable() {
-  return <div className="w-full h-4 animate-pulse">Loading....</div>;
+  return <div className="space-y-4">Loading</div>;
 }
